Fix Display

+++ resources/views/admin/programs/index.blade.php
@@ -12,11 +12,13 @@
                                 <tr>
                                     <th>#</th>
                                     <th>{{ __('Name Program') }}</th>
-                                    <th>
-                                        <a href="{{ route('programs.create') }}" class="btn m-btn--pill m-btn--air btn-secondary table-middle " data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Add New') }}">
-                                            <i class="flaticon-add"></i>
-                                        </a>
-                                    </th>
+                                    @if (Auth::user()->role == config('site.permission.admin'))
+                                        <th>
+                                            <a href="{{ route('programs.create') }}" class="btn m-btn--pill m-btn--air btn-secondary table-middle " data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Add New') }}">
+                                                <i class="flaticon-add"></i>
+                                            </a>
+                                        </th>
+                                    @endif
                                 </tr>
                             </thead>
                             <tbody>
@@ -24,12 +26,16 @@
                                 <tr>
                                     <th scope="row">{{ $key + 1}}</th>
                                     <td>{{ $pro->name }}</td>
-                                    <td><a href="{{ route('programs.edit', ['id' => $pro->id]) }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-skin="dark" data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Edit') }}">
-                                    <i class="flaticon-edit-1"></i>
-                                    </a>
-                                    {!! Form::open(['route' => ['programs.destroy', $pro['id'] ],'class' => 'd-inline', 'method' => 'delete']) !!}
-                                        {!! Form::button('<i class="flaticon-cancel"></i>', ['class' => 'btn btn-danger m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill delete', 'data-toggle' => 'm-tooltip', 'data-placement' => 'right','data-skin' => 'dark', 'data-original-title' => __('Delete'), 'type' => 'submit' ]) !!}
-                                    {!! Form::close() !!}</td>
+                                    @if (Auth::user()->role == config('site.permission.admin'))
+                                        <td>
+                                            <a href="{{ route('programs.edit', ['id' => $pro->id]) }}" class="btn btn-warning m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill" data-skin="dark" data-toggle="m-tooltip" data-placement="left" data-original-title="{{ __('Edit') }}">
+                                        <i class="flaticon-edit-1"></i>
+                                        </a>
+                                        {!! Form::open(['route' => ['programs.destroy', $pro['id'] ],'class' => 'd-inline', 'method' => 'delete']) !!}
+                                            {!! Form::button('<i class="flaticon-cancel"></i>', ['class' => 'btn btn-danger m-btn m-btn--icon m-btn--icon-only m-btn--custom m-btn--pill delete', 'data-toggle' => 'm-tooltip', 'data-placement' => 'right','data-skin' => 'dark', 'data-original-title' => __('Delete'), 'type' => 'submit' ]) !!}
+                                        {!! Form::close() !!}
+                                        </td>
+                                    @endif
                                 </tr>
                                 @endforeach
                             </tbody>
